requirement.txt
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
decorator==5.1.1
future==0.18.3
geocoder==1.38.1
idna==3.4
python-dateutil==2.8.2
ratelim==0.1.6
requests==2.31.0
six==1.16.0
urllib3==2.0.4
